+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed:
 - Declutter task form
+- Simplified editor toolbar
 
 ## [0.8.225] - 2023-01-06
 ### Changed: