+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed:
 - Declutter task form
 - Simplified editor toolbar
+- Unified searchable list
 
 ## [0.8.225] - 2023-01-06
 ### Changed: