+++ CHANGELOG.md
@@ -6,13 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed:
+- New search header in journal
+- Upgraded dependencies
+
+## [0.8.226] - 2023-01-07
+### Changed:
 - Declutter task form
 - Simplified editor toolbar
 - Unified searchable list
 - Hide task title label when task defined
 - Floating search bar replaced in settings
 - Floating search bar replaced in journal
-- New search header in journal
 
 ## [0.8.225] - 2023-01-06
 ### Changed: